[Feat] Pay 결제 시 계좌 입금 & 출금 요청

+++ src/main/java/com/click/payment/domain/dto/request/PaymentHistoryRequest.java
@@ -7,7 +7,7 @@ import java.time.LocalDateTime;
 
 public record PaymentHistoryRequest(
     Long payNum,
-    Integer payAmount,
+    Long payAmount,
     String failRedirUrl,
     String successRedirUrl
 ) {
